density-size-correction/density-corrected-natural-size.html

<!DOCTYPE html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="author" title="Noam Rosenthal" href="noam@webkit.org">
<img src="resources/exif-resolution-invalid-cm.jpg" data-width="100" data-height="50" class="test-img" />
<img src="resources/exif-resolution-invalid-no-match.jpg" data-width="100" data-height="50" class="test-img" />
<img src="resources/exif-resolution-invalid-partial.jpg" data-width="100" data-height="50" class="test-img" />
<img src="resources/exif-resolution-no-change.jpg" data-width="100" data-height="50" class="test-img" />
<img src="resources/exif-resolution-none.jpg" data-width="100" data-height="50" class="test-img" />
<img src="resources/exif-resolution-valid-hires.jpg" data-width="50" data-height="25" class="test-img" />
<img src="resources/exif-resolution-valid-lores.jpg" data-width="200" data-height="100" class="test-img" />
<img src="resources/exif-resolution-valid-non-uniform.jpg" data-width="50" data-height="100" class="test-img" />
<img src="resources/exif-resolution-with-orientation.jpg" data-width="100" data-height="200" class="test-img" />
<script>
    let run_test = () => {
        test(() => {
            [...document.querySelectorAll('.test-img')].forEach(img => {
                assert_equals(img.naturalWidth, +img.dataset.width, `Density size correction for ${img.src}`);
                assert_equals(img.naturalHeight, +img.dataset.height, `Density size correction for ${img.src}`);
            })
        }, "Test the image dimensions of different EXIF scenarios");
    }
    window.addEventListener("load", run_test);
</script>

</body>
</html>
